+++ html2text/cli.py
@@ -24,6 +24,13 @@ def main():
         version='%prog ' + ".".join(map(str, __version__))
     )
     p.add_option(
+        "--pad-tables",
+        dest="pad_tables",
+        action="store_true",
+        default=config.PAD_TABLES,
+        help="pad the cells to equal column width in tables"
+    )
+    p.add_option(
         "--no-wrap-links",
         dest="wrap_links",
         action="store_false",
@@ -271,5 +278,6 @@ def main():
     h.links_each_paragraph = options.links_each_paragraph
     h.mark_code = options.mark_code
     h.wrap_links = options.wrap_links
+    h.pad_tables = options.pad_tables
 
     wrapwrite(h.handle(data))