+++ test/test_html2text.py
@@ -170,6 +170,10 @@ def generate_test(fn):
         module_args['mark_code'] = True
         cmdline_args.append('--mark-code')
 
+    if base_fn.startswith('pad_table'):
+        module_args['pad_tables'] = True
+        cmdline_args.append('--pad-tables')
+
     if base_fn not in ['bodywidth_newline.html', 'abbr_tag.html']:
         test_func = None